+++ app/src/components/drawer/index.tsx
@@ -10,13 +10,14 @@ import {
   TouchableWithoutFeedback,
   Alert,
   useWindowDimensions,
+  Text,
+  TouchableOpacity,
 } from "react-native";
 import Modal from "react-native-modal";
 import DrawerItem from "./drawer-item";
 import LegalItem from "./legal-item";
 import localStorage from "../../utils/localStorage";
 import { getBadgeNotesVersion } from "../../scenes/news";
-import Text from "../../components/MyText";
 import { colors } from "../../utils/colors";
 import NeedUpdateContext from "../../context/needUpdate";
 import { recommendApp } from "../../utils/share";
@@ -33,6 +34,14 @@ import LightBulb from "@assets/svg/icon/LightBulb";
 import Download from "@assets/svg/icon/Download";
 import NPS from "../../services/NPS/NPS";
 import Gear from "@assets/svg/Gear";
+import { SquircleButton, SquircleView } from "expo-squircle-view";
+import { mergeClassNames } from "@/utils/className";
+import { typography } from "@/utils/typography";
+import ChevronIcon from "@assets/svg/icon/chevron";
+import MegaphoneIcon from "@assets/svg/icon/MegaphoneIcon";
+import { TW_COLORS } from "@/utils/constants";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import ArrowRightSvg from "../../../assets/svg/arrow-right.js";
 
 export default ({ navigation, visible, onClick }) => {
   const [isVisible, setIsVisible] = useState();
@@ -45,6 +54,7 @@ export default ({ navigation, visible, onClick }) => {
   const [badgeNpsProIsVisible, setBadgeNpsProIsVisible] = useState(false);
   const [badgeNotesVersionVisible, setBadgeNotesVersionVisible] = useState(false);
   const [NPSvisible, setNPSvisible] = useState(false);
+  const insets = useSafeAreaInsets();
 
   useEffect(() => {
     setIsVisible(visible);
@@ -89,83 +99,156 @@ export default ({ navigation, visible, onClick }) => {
         deviceWidth={width}
         deviceHeight={height} // <--- met à jour automatiquement
       >
-        <SafeAreaView style={[styles.card]}>
-          <ScrollView
-            showsVerticalScrollIndicator={false}
-            contentContainerStyle={styles.scrollContainer}
-            style={[
-              // styles.card,
-              {
-                // on android the scrollview is being desactived when you open several time the drawer
-                // apparently it is a bug when having scrollview inside modal
-                // this weird workaround fixes it https://github.com/facebook/react-native/issues/48822#issuecomment-2667011212
-                height: 100,
-              },
-            ]}
-          >
-            <Text style={styles.title}>Jardin Mental</Text>
-            <DrawerItem badge={badgeNotesVersionVisible} title="Nouveautés" path="news" navigation={navigation} onClick={onClick} icon={<Star />} />
-            <Separator />
-            <DrawerItem title="Comment ça marche ?" path="faq" navigation={navigation} onClick={onClick} icon={<StickerSquare />} />
-            <DrawerItem title="Recommander Jardin&nbsp;Mental" onClick={recommendApp} navigation={navigation} icon={<Share />} />
-            <DrawerItem title="Parler à quelqu'un et s'informer" path="infos" navigation={navigation} onClick={onClick} icon={<Phone />} />
-            <DrawerItem title="Nous contacter" path="contact" navigation={navigation} onClick={onClick} icon={<MessageTextCircle />} />
-            <DrawerItem title="Qui peut voir mes données ?" path="privacy-light" navigation={navigation} onClick={onClick} icon={<Lock />} />
-            {updateVisible ? (
-              <DrawerItem
-                badge
-                title="Mettre à jour"
-                icon={<Bell />}
-                navigation={navigation}
-                onClick={() =>
-                  Linking.openURL(
-                    Platform.OS === "ios"
-                      ? "itms-apps://apps.apple.com/FR/app/id1540061393"
-                      : "https://play.app.goo.gl/?link=https://play.google.com/store/apps/details?id=com.monsuivipsy"
-                  )
-                }
-              />
-            ) : null}
-            <DrawerItem
-              badge={badgeNpsProIsVisible}
-              title="Donner mon avis"
-              icon={<LightBulb />}
-              navigation={navigation}
-              onClick={async () => {
-                // dismiss the drawer first, IOS cannot display two modals at the same tme
-                setIsVisible(false);
-                await localStorage.setVisitProNPS(true);
-                setTimeout(() => {
-                  // a bit hacky : whai for drawer to dismiss before displaying NPS
-                  setNPSvisible(true);
-                }, 500);
-              }}
-            />
-            {isDevMode && <DrawerItem title="Dev Mode" path="dev-mode" navigation={navigation} onClick={onClick} icon={<Gear />} />}
-            {isDevMode && (
-              <DrawerItem
-                title="Export / Import mes données"
-                path="data-export-import"
-                navigation={navigation}
-                onClick={onClick}
-                icon={<Download />}
-              />
-            )}
-            <Separator />
-            <LegalItem title="Conditions générales d'utilisation" path="cgu" navigation={navigation} onClick={onClick} />
-            <LegalItem title="Politique de confidentialité" path="privacy" navigation={navigation} onClick={onClick} />
-            <LegalItem title="Mentions légales" path="legal-mentions" navigation={navigation} onClick={onClick} />
-            <TouchableWithoutFeedback onPress={handleDevModePress}>
-              <View style={styles.versionContainer}>
-                <Text style={styles.versionLabel}>
-                  {Platform.OS === "ios"
-                    ? `${app.expo.version} (${app.expo.ios.buildNumber})`
-                    : `${app.expo.version} (${app.expo.android.versionCode})`}
-                </Text>
+        <SquircleView cornerSmoothing={100} preserveSmoothing={true} style={[styles.card]}>
+          <View className="bg-cnam-primary-800" style={{ width: "100%", height: "100%" }}>
+            <ScrollView
+              showsVerticalScrollIndicator={false}
+              contentContainerStyle={styles.scrollContainer}
+              style={[
+                // styles.card,
+                {
+                  // on android the scrollview is being desactived when you open several time the drawer
+                  // apparently it is a bug when having scrollview inside modal
+                  // this weird workaround fixes it https://github.com/facebook/react-native/issues/48822#issuecomment-2667011212
+                  height: 100,
+                },
+              ]}
+            >
+              <View
+                className="bg-cnam-primary-800"
+                style={{
+                  paddingTop: insets.top,
+                }}
+              >
+                <View className="p-4 pb-6 bg-cnam-primary-800 flex-col space-y-4">
+                  <Text className={mergeClassNames(typography.displayXsBold, "text-cnam-primary-25 text-left")}>Jardin Mental</Text>
+                  <SquircleButton
+                    onPress={() => {
+                      navigation.navigate("news");
+                      onClick();
+                    }}
+                    className="bg-cnam-primary-900 px-3 py-2 flex-row self-start items-center"
+                    cornerSmoothing={100}
+                    preserveSmoothing={true}
+                    style={{
+                      borderRadius: 16,
+                    }}
+                  >
+                    {badgeNotesVersionVisible && <View className="bg-red rounded-full w-2 h-2 mr-1"></View>}
+                    <Text className={mergeClassNames(typography.textMdMedium, "text-cnam-primary-25")}>Nouveautés</Text>
+                    <View
+                      className="ml-2"
+                      style={{
+                        top: 1,
+                        transform: [{ scaleX: -1 }],
+                      }}
+                    >
+                      <ChevronIcon strokeWidth={2.5} color={TW_COLORS.CNAM_PRIMARY_25} />
+                    </View>
+                  </SquircleButton>
+                </View>
+                <TouchableOpacity
+                  onPress={async () => {
+                    // dismiss the drawer first, IOS cannot display two modals at the same tme
+                    onClick();
+                    await localStorage.setVisitProNPS(true);
+                    setTimeout(() => {
+                      // a bit hacky : whai for drawer to dismiss before displaying NPS
+                      setNPSvisible(true);
+                    }, 500);
+                  }}
+                  className="bg-white rounded-t-xl flex-row px-4 py-6 self-center border-gray-300 border-b-0 justify-between items-center relative"
+                  style={{ width: "90%" }}
+                >
+                  <View className="flex-row items-center space-x-4">
+                    <View className="border border-[1.5] border-cnam-primary-800 rounded-full w-[32] h-[32] items-center justify-center">
+                      <MegaphoneIcon color={TW_COLORS.CNAM_PRIMARY_800} />
+                    </View>
+                    <Text className={mergeClassNames(typography.textLgSemibold, "text-cnam-primary-950")}>Donner mon avis</Text>
+                  </View>
+                  <ArrowRightSvg color={TW_COLORS.GRAY_600} />
+                </TouchableOpacity>
               </View>
-            </TouchableWithoutFeedback>
-          </ScrollView>
-        </SafeAreaView>
+              <View className="bg-cnam-primary-50">
+                <TouchableOpacity
+                  onPress={async () => {
+                    onClick();
+                    navigation.navigate("contact");
+                  }}
+                  className="bg-white rounded-b-xl flex-row px-4 py-6 self-center border border-gray-300 border-t-0 justify-between items-center"
+                  style={{ width: "90%" }}
+                >
+                  <View
+                    className="absolute -top-[1] bg-gray-400 h-[1]"
+                    style={{
+                      left: "5%",
+                      right: "5%",
+                    }}
+                  ></View>
+                  <View className="flex-row items-center space-x-4">
+                    <View className="border border-[1.5] border-cnam-primary-800 rounded-full w-[32] h-[32] items-center justify-center">
+                      <MessageTextCircle color={TW_COLORS.CNAM_PRIMARY_800} />
+                    </View>
+                    <Text className={mergeClassNames(typography.textLgSemibold, "text-cnam-primary-950")}>Nous contacter</Text>
+                  </View>
+                  <ArrowRightSvg color={TW_COLORS.GRAY_600} />
+                </TouchableOpacity>
+                <View className="mt-6 mb-8">
+                  <DrawerItem title="Comment ça marche ?" path="faq" navigation={navigation} onClick={onClick} />
+                  <Separator />
+                  <DrawerItem title="Qui peut voir mes données ?" path="privacy-light" navigation={navigation} onClick={onClick} />
+                  <Separator />
+                  <DrawerItem title="Recommander l'app" onClick={recommendApp} navigation={navigation} />
+                  <Separator />
+                  <DrawerItem
+                    style={{
+                      backgroundColor: TW_COLORS.CNAM_CYAN_100_LIGHTEN_80,
+                    }}
+                    title="Contacts utiles 24h/24 - 7J/7"
+                    path="infos"
+                    navigation={navigation}
+                    onClick={onClick}
+                  />
+                  <Separator />
+                  {updateVisible ? (
+                    <DrawerItem
+                      badge
+                      title="Mettre à jour"
+                      icon={<Bell />}
+                      navigation={navigation}
+                      onClick={() =>
+                        Linking.openURL(
+                          Platform.OS === "ios"
+                            ? "itms-apps://apps.apple.com/FR/app/id1540061393"
+                            : "https://play.app.goo.gl/?link=https://play.google.com/store/apps/details?id=com.monsuivipsy"
+                        )
+                      }
+                    />
+                  ) : null}
+                  {isDevMode && (
+                    <>
+                      <DrawerItem title="Dev Mode" path="dev-mode" navigation={navigation} onClick={onClick} />
+                      <Separator />
+                      <DrawerItem title="Export / Import mes données" path="data-export-import" navigation={navigation} onClick={onClick} />
+                    </>
+                  )}
+                </View>
+                <LegalItem title="Conditions générales d'utilisation" path="cgu" navigation={navigation} onClick={onClick} />
+                <LegalItem title="Politique de confidentialité" path="privacy" navigation={navigation} onClick={onClick} />
+                <LegalItem title="Mentions légales" path="legal-mentions" navigation={navigation} onClick={onClick} />
+                <TouchableWithoutFeedback onPress={handleDevModePress}>
+                  <View style={styles.versionContainer}>
+                    <Text style={styles.versionLabel}>
+                      {Platform.OS === "ios"
+                        ? `${app.expo.version} (${app.expo.ios.buildNumber})`
+                        : `${app.expo.version} (${app.expo.android.versionCode})`}
+                    </Text>
+                  </View>
+                </TouchableWithoutFeedback>
+              </View>
+            </ScrollView>
+          </View>
+        </SquircleView>
       </Modal>
     </>
   );
@@ -179,13 +262,12 @@ const styles = StyleSheet.create({
     flex: 1,
   },
   separator: {
-    borderColor: "#eee",
+    borderColor: TW_COLORS.CNAM_PRIMARY_500,
     borderTopWidth: 1,
-    marginHorizontal: 30,
-    marginVertical: 15,
   },
   scrollContainer: {
     paddingBottom: 80,
+    backgroundColor: TW_COLORS.CNAM_PRIMARY_50,
   },
   versionContainer: {
     marginTop: 47,
@@ -201,11 +283,10 @@ const styles = StyleSheet.create({
   card: {
     width: "80%",
     height: "100%",
-    backgroundColor: "#fff",
-    borderRadius: 10,
+    borderRadius: 16,
     borderTopLeftRadius: 0,
     borderBottomLeftRadius: 0,
-    paddingBottom: 30,
+    overflow: "hidden",
   },
   title: {
     fontSize: 20,