+++ app/src/scenes/dev-mode/index.tsx
@@ -30,7 +30,7 @@ const CollapsibleSection = ({ title, children }) => {
 
 const DevMode = ({ navigation }) => {
   const { clearProfile, loadProfile } = useUserProfile();
-  const [_diaryData, _addEntryToDiaryData, internal__deleteDiaryData, importDiaryData] = useContext(DiaryDataContext);
+  const [_diaryData, _addEntryToDiaryData, internal__deleteDiaryData, importDiaryData, internal__deleteAllDiaryData] = useContext(DiaryDataContext);
   const disableDevMode = async () => {
     await AsyncStorage.setItem("devMode", "false");
     navigation.navigate("tabs");
@@ -106,16 +106,22 @@ const DevMode = ({ navigation }) => {
             confirmText: "Confirmer",
             cancelText: "Annuler",
             onConfirm: async () => {
-              await wipeData();
-              await AsyncStorage.clear();
-              await clearProfile();
-              await loadProfile();
-              navigation.reset({
-                index: 0,
-                routes: [{ name: "onboarding" }],
-              });
+              try {
+                await internal__deleteAllDiaryData();
+                await wipeData();
+                await AsyncStorage.clear();
+                await clearProfile();
+                await loadProfile();
+                navigation.reset({
+                  index: 0,
+                  routes: [{ name: "onboarding" }],
+                });
+              } catch (e) {
+                console.error(e);
+              }
             },
-            onCancel: () => {
+            onCancel: (e) => {
+              console.log(e);
               // No action needed, dialog will close automatically
             },
           });