+++ app/src/scenes/onboarding-v2/BannerHeader.tsx
@@ -83,9 +83,9 @@ export default function BannerHeader({
           animatedStatusBarColor,
           hidden ? { opacity: 0 } : undefined,
         ]}
-        className={`${(SHARED_HEADER || hideHeader) && !HEADER_WITH_BANNER ? "pt-16" : ""}`}
+        className={`${hideHeader ? "pt-16" : ""}`}
       >
-        {(!(SHARED_HEADER || hideHeader) || HEADER_WITH_BANNER) && (
+        {!hideHeader && (
           <CheckInHeader
             title={headerTitle || ""}
             dynamicTitle={dynamicTitle}